Store payloads in a queue when arriving in parallel
When messages arrive very quickly after each other, or the message handler is waiting for the Blob arraybuffer to resolve (it's async), it can happen that another message that arrived in parallel resolves the promise already, loosing the first message.

This commit introduces a payload queue, in which payloads that take longer are stored, and taken out on the next iteration.

Attention: The queue is only processed when a new message arrives.